This patch activates the middle mouse button on my Synaptics TouchPad (installed in a Dell M6500). Not only does the patch enable a functional middle mouse button, it also fixes the strange "right-click mouse button + TouchPad scroll bug" discussed in at least a couple of places on the internet (including a post by me back in 2010):
http://forums.freebsd.org/showthread.php?t=12941
http://forums.freebsd.org/showthread.php?t=10125
http://lists.freebsd.org/pipermail/freebsd-mobile/2010-February/011929.html

Although I can't test the additional features this patch adds, e.g.
extended button support, due to hardware unavailability, I can verify
that the middle button works and the right-click spurious scroll issue
is resolved.

Also, the Synaptics PS/2 TouchPad Interfacing Guide formalizes the
support of "guest devices", which was previously represented by a
reserved bit. See section 5 of the guide:

http://www.synaptics.com/sites/default/files/511-000275-01_RevB.pdf

One important note to this is the dropped support for guest devices
from the X.org synaptics driver as of this commit:

http://cgit.freedesktop.org/xorg/driver/xf86-input-synaptics/commit/?id=b19e3782a77c171ca20fc962f95923495fdb7978

The guest device in my Dell M6500 is the Synaptics TouchStyk, an
enhanced "trackpoint"-style pointing device, which theoretically could
be represented as another pointing device (similar to how the Linux
kernel handles these devices from what I've read). However, the
Synaptics support provided by the psm(4) driver needs significantly
more work to support the devices in this manner, so this patch is more
of a stop-gap for now.